+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix support for keys in new queue handler.
 - Fix the bug where the query itself changes after encoding
 - Fix the dependency error in copy_vectors within vector_index. 
+- Fix Template substitutions
 
 
 ## [0.2.0](https://github.com/superduper-io/superduper/compare/0.1.3...0.2.0])    (2024-Jun-21)